requirements.txt
aioredis==2.0.1
python-dotenv==1.0.1
fastapi==0.115.12
httpx==0.28.1
mcp==1.12.3
numpy==1.26.4
openai==1.77.0
pandas==2.2.3
pydantic==2.11.4
requests==2.32.5
shortuuid==1.0.13
tqdm==4.67.1
uvicorn==0.34.2
websockets==15.0.1
aiofiles==24.1.0
colorama==0.4.6
uv==0.6.9
elasticsearch[async]==7.13.0
msgpack==1.1.0
aiohttp==3.11.18
aiohttp-sse-client==0.2.1
python-multipart==0.0.20
pillow==11.2.1
respx==0.22.0
aioresponses==0.7.8
dashscope
PyMuPDF
python-pptx
openpyxl